Rating Stuff :(

+++ src/MusicApp2017/Migrations/20170420005610_Add-Migration Rating.cs
@@ -16,7 +16,7 @@ namespace MusicApp2017.Migrations
                     RatingID = table.Column<int>(nullable: false)
                         .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
                     AlbumID = table.Column<int>(nullable: false),
-                    Score = table.Column<int>(nullable: false)
+                    Score = table.Column<decimal>(nullable: false)
                 },
                 constraints: table =>
                 {
